requirements.txt
openai-whisper
pretty_midi
transformers
moviepy
